fix: separate unit and integration tests by folders
