+++ src/tests/integration/app.test.ts
@@ -1,14 +1,16 @@
 import request from 'supertest';
-import server from '../app';
+import server from '../../app';
 describe('Get Endpoint /', () => {
+
+  
   it('Get /', async () => {
-    const res =  await  request(server)
+    const res = await request(server)
       .get('/');
     expect(res.statusCode).toEqual(200);
     expect(res.body).toHaveProperty('name');
     expect(res.body).toHaveProperty('version');
   });
-  afterEach(done  => {
+  afterEach(done => {
     // close server conection
     server.close();
     done();
