+++ src/tests/integration/cyclistRoutes.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import server from '../app';
+import server from '../../app';
 
 describe('Get one Cyclist', () => {
   it('Get', async () => {
@@ -26,7 +26,7 @@ describe('Get one Cyclist', () => {
 });
 
 
-describe('Create one cyclist', ()=>{
+describe('Create one cyclist', () => {
   it('Create a cyclist', async () => {
     const cyclist = {
       'name': 'string',
@@ -65,7 +65,7 @@ describe('Create one cyclist', ()=>{
   });
 });
 
-describe('Create a invalid cyclist', ()=>{
+describe('Create a invalid cyclist', () => {
   it('Create a cyclist with missing params', async () => {
     const cyclist = {
       'passaporte': {
@@ -91,7 +91,7 @@ describe('Create a invalid cyclist', ()=>{
     const cyclist = '';
 
     const res = await request(server)
-      .post('/cyclist').send(cyclist );
+      .post('/cyclist').send(cyclist);
 
     expect(res.statusCode).toEqual(400);
     expect(res.body).toHaveProperty('error');
